feat : user page

+++ src/com/booklink/ui/panel/content/book/BookContentPanel.java
@@ -11,6 +11,7 @@ import com.booklink.ui.panel.content.book.booksummary.BookSummaryPanel;
 import javax.swing.*;
 import java.awt.*;
 import java.util.List;
+import java.util.Set;
 // 1488, 970
 
 
@@ -38,7 +39,7 @@ public class BookContentPanel extends ContentPanel {
         bookController = new BookController();
         // books And count를 가지고 온다.
         books = bookController.findAllBookWithCount();
-        maxPage = Math.max(1, (int) Math.ceil(books.size() / pagePerContent));
+        maxPage = Math.max(1, (int) Math.ceil((double) books.size() / pagePerContent));
         currentPage = 1;
 
         // ContentPanel에서 시작하는 번호와, 끝번호를 가지고 있어야 한다.
@@ -51,7 +52,7 @@ public class BookContentPanel extends ContentPanel {
         bookController = new BookController();
         // books And count를 가지고 온다.
         books = bookController.findBooksByContainsTitle(title);
-        maxPage = Math.max(1, (int) Math.ceil(books.size() / pagePerContent));
+        maxPage = Math.max(1, (int) Math.ceil((double) books.size() / pagePerContent));
         currentPage = 1;
 
         // ContentPanel에서 시작하는 번호와, 끝번호를 가지고 있어야 한다.
@@ -72,6 +73,8 @@ public class BookContentPanel extends ContentPanel {
         update(currentPage);
 
     }
+
+
 
     // update가 호출되면 pageNum를 갱신하고 해당 페이지로 이동시킨다.
     @Override
